experiments/pyproject.toml
[project]
name = "create_db"
version = "0.1.0"
description = "Minimal project: dipendenza neo4j per caricare il DB sul grafo"
requires-python = ">=3.9"
dependencies = [
    "neo4j>=5.0.0",
    "polars>=0.27.0",
    "numpy>=1.24.0",
    "hopwise",
    "hopwise[pathlm]",
    "hydra-core>=1.3.2",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"